fixe mag data processing and added python program to visualise data correction in real time

+++ src/BiasMatrix.h
@@ -3,7 +3,7 @@
 
 struct BiasMatrix
 {
-    float biasMatrix[3];
+    double biasMatrix[3];
 };
 extern BiasMatrix globalBiasMatrix;
 #endif
