+++ src/ScaleMatrix.h
@@ -3,7 +3,7 @@
 
 struct ScaleMatrix
 {
-    float scaleMatrix[3][3];
+    double scaleMatrix[3][3];
 };
 extern ScaleMatrix globalScaleMatrix;
